+++ packages/backend/src/tokens/types.ts
@@ -4,6 +4,9 @@ import type { ApiClient, APIKey, IdPOAuthAccessToken, MachineToken } from '../ap
 import type { TokenType } from './tokenTypes';
 import type { VerifyTokenOptions } from './verify';
 
+/**
+ * @interface
+ */
 export type AuthenticateRequestOptions = {
   /**
    * The Clerk Publishable Key from the [**API keys**](https://dashboard.clerk.com/last-active?path=api-keys) page in the Clerk Dashboard.
@@ -58,7 +61,7 @@ export type AuthenticateRequestOptions = {
 } & VerifyTokenOptions;
 
 /**
- * @expand
+ * @inline
  */
 export type OrganizationSyncOptions = {
   /**
@@ -68,8 +71,7 @@ export type OrganizationSyncOptions = {
    *
    * Patterns must have a path parameter named either `:id` (to match a Clerk organization ID) or `:slug` (to match a Clerk organization slug).
    *
-   * > [!WARNING]
-   * > If the organization can't be activated—either because it doesn't exist or the user lacks access—the previously active organization will remain unchanged. Components must detect this case and provide an appropriate error and/or resolution pathway, such as calling `notFound()` or displaying an [`<OrganizationSwitcher />`](https://clerk.com/docs/components/organization/organization-switcher).
+   * If the organization can't be activated—either because it doesn't exist or the user lacks access—the previously active organization will remain unchanged. Components must detect this case and provide an appropriate error and/or resolution pathway, such as calling `notFound()` or displaying an [`<OrganizationSwitcher />`](https://clerk.com/docs/components/organization/organization-switcher).
    *
    * @example
    * ["/orgs/:slug", "/orgs/:slug/(.*)"]
@@ -81,7 +83,7 @@ export type OrganizationSyncOptions = {
   organizationPatterns?: Pattern[];
 
   /**
-   * URL patterns for resources that exist within the context of a [Clerk Personal Account](https://clerk.com/docs/organizations/organization-workspaces#organization-workspaces-in-the-clerk-dashboard:~:text=Personal%20account) (user-specific, outside any organization).
+   * URL patterns for resources that exist within the context of a [Clerk Personal Account](https://clerk.com/docs/organizations/organization-workspaces) (user-specific, outside any organization).
    *
    * If the route also matches the `organizationPattern` prop, the `organizationPattern` prop takes precedence.
    *