+++ packages/clerk-js/src/ui/components/OrganizationProfile/OrganizationMembersTabRequests.tsx
@@ -1,7 +1,9 @@
+import { Header } from '@/ui/elements/Header';
+import { ProfileSection } from '@/ui/elements/Section';
+
 import { Protect } from '../../common';
 import { useEnvironment, useOrganizationProfileContext } from '../../contexts';
 import { Col, Flex, localizationKeys, Text } from '../../customizables';
-import { Header, ProfileSection } from '../../elements';
 import { mqu } from '../../styledSystem';
 import { DomainList } from './DomainList';
 import { MembersActionsRow } from './MembersActions';