+++ packages/clerk-js/src/ui/components/SignIn/__tests__/SignInStart.test.tsx
@@ -2,11 +2,12 @@ import { OAUTH_PROVIDERS } from '@clerk/shared/oauth';
 import type { SignInResource } from '@clerk/types';
 import { waitFor } from '@testing-library/react';
 
+import { CardStateProvider } from '@/ui/elements/contexts';
+
 import { ClerkAPIResponseError } from '../../../../core/resources';
 import { fireEvent, mockWebAuthn, render, screen } from '../../../../testUtils';
 import { OptionsProvider } from '../../../contexts';
 import { AppearanceProvider } from '../../../customizables';
-import { CardStateProvider } from '../../../elements';
 import { bindCreateFixtures } from '../../../utils/test/createFixtures';
 import { SignInStart } from '../SignInStart';