+++ packages/clerk-js/src/ui/components/UserProfile/index.tsx
@@ -1,10 +1,13 @@
 import type { UserProfileModalProps, UserProfileProps } from '@clerk/types';
 import React from 'react';
 
+import { withCardStateProvider } from '@/ui/elements/contexts';
+import { NavbarMenuButtonRow } from '@/ui/elements/Navbar';
+import { ProfileCard } from '@/ui/elements/ProfileCard';
+
 import { USER_PROFILE_CARD_SCROLLBOX_ID } from '../../constants';
 import { UserProfileContext, withCoreUserGuard } from '../../contexts';
 import { Flow, localizationKeys } from '../../customizables';
-import { NavbarMenuButtonRow, ProfileCard, withCardStateProvider } from '../../elements';
 import { Route, Switch } from '../../router';
 import type { UserProfileCtx } from '../../types';
 import { UserProfileNavbar } from './UserProfileNavbar';